refactor: 매번 allNum 생성하지 않도록 변경
